requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7
python-dotenv==1.0.0
email-validator==2.0.0
boto3==1.34.34
psycopg2-binary==2.9.9
python-jose==3.3.0
requests==2.31.0
watchtower==3.0.1  # For CloudWatch logging 